+++ deadgrep.el
@@ -149,6 +149,13 @@ display."
 (defvar-local deadgrep--file-type 'all)
 (put 'deadgrep--file-type 'permanent-local t)
 
+(defvar-local deadgrep--skip-if-hidden nil
+  "Whether deadgrep should ignore hidden files (e.g. files called .foo).")
+(put 'deadgrep--skip-if-hidden 'permanent-local t)
+(defvar-local deadgrep--skip-if-vcs-ignore 't
+  "Whether deadgrep should ignore files if they're listed in .gitignore.")
+(put 'deadgrep--skip-if-vcs-ignore 'permanent-local t)
+
 (defvar-local deadgrep--context nil
   "When set, also show context of results.
 This is stored as a cons cell of integers (lines-before . lines-after).")
@@ -497,6 +504,24 @@ with a text face property `deadgrep-match-face'."
   'case nil
   'help-echo "Change file type")
 
+(define-button-type 'deadgrep-skip-hidden-type
+  'action #'deadgrep--skip-if-hidden
+  'case nil
+  'help-echo "Toggle whether to skip dotfiles")
+
+(defun deadgrep--skip-if-hidden (_button)
+  (setq deadgrep--skip-if-hidden (not deadgrep--skip-if-hidden))
+  (deadgrep-restart))
+
+(define-button-type 'deadgrep-vcs-skip-type
+  'action #'deadgrep--skip-if-vcs-ignore
+  'case nil
+  'help-echo "Toggle whether to skip files listed in .gitignore")
+
+(defun deadgrep--skip-if-vcs-ignore (_button)
+  (setq deadgrep--skip-if-vcs-ignore (not deadgrep--skip-if-vcs-ignore))
+  (deadgrep-restart))
+
 (defun deadgrep--format-file-type (file-type extensions)
   (let* ((max-exts 4)
          (truncated (> (length extensions) max-exts)))
@@ -736,6 +761,11 @@ to obtain ripgrep results."
       (push (format "--before-context=%s" (car context)) args)
       (push (format "--after-context=%s" (cdr context)) args))
 
+    (unless deadgrep--skip-if-hidden
+      (push "--hidden" args))
+    (unless deadgrep--skip-if-vcs-ignore
+      (push "--no-ignore-vcs" args))
+
     (push "--" args)
     (push search-term args)
     (push "." args)
@@ -839,6 +869,15 @@ search settings."
             (if (eq (car-safe deadgrep--file-type) 'glob)
                 (format ":%s" (cdr deadgrep--file-type))
               "")
+            "\n"
+            (propertize "Skip: "
+                        'face 'deadgrep-meta-face)
+            (deadgrep--button "dotfiles" 'deadgrep-skip-hidden-type)
+            (if deadgrep--skip-if-hidden ":yes" ":no")
+            " "
+            (deadgrep--button ".gitignore items" 'deadgrep-vcs-skip-type)
+            (if deadgrep--skip-if-vcs-ignore ":yes" ":no")
+            " "
             "\n\n")
     (put-text-property
      start-pos (point)